Move browser_action to a tool file (#2099)
